+++ src/main/java/com/coherentsolutions/pot/insurance/service/EmployeeService.java
@@ -2,7 +2,9 @@ package com.coherentsolutions.pot.insurance.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
+/*Employer Portal is used to manage company and user profiles, enroll users into insurance plans,
+as well as to view enrollments and file claims – all depending in user permissions/functions.
+*/
 @Service
 @RequiredArgsConstructor
 public class EmployeeService {
